perf(pool): drop register_usage from the parse hot path
register_usage built a container->child back-reference index on EVERY parse
(a map lookup + a shared_ptr copy per child of every container) that nothing
read — find_containers_of is unused by save/loader/Python; only its own unit
tests. Removed the calls from Frame::finish and SharedPtrBuilder; the manual
mechanism + its direct tests remain. Ibex parse+RT 95.4s -> 90.1s (~5%); frees
per-child overhead + the index's memory. Shadow still 0 divergences.

+++ src/grammar.cpp
@@ -2417,9 +2417,9 @@ tl::expected<ValuePtr, ParseError> Grammar::parse_from(
                 default: return "scalar";
                 }
             };
-            std::fprintf(stderr, "diverged start=%u depth=%zu shadow=%s frame=%s\n",
-                start.value(), shadow.depth(), shape(shadow_result).c_str(),
-                shape(result_value).c_str());
+            std::fprintf(stderr, "diverged start=%zu depth=%zu shadow=%s frame=%s\n",
+                static_cast<std::size_t>(start.value()), shadow.depth(),
+                shape(shadow_result).c_str(), shape(result_value).c_str());
             shadow_first_diff(shadow_result, result_value, "");
         }
     }
